Update SD class and add new functions
Update the SD class so it now will always take a
function to calculate the gatekeeping threshold
and lambda values. Split the lambdas_function
into two functions, proportional_gatekeeping and
fixed_gatekeeping. These are hard-coded functions
that are used to calculate the lambda values and
are provided as input to the SD class.